Fix casts for WASM build

+++ CesiumGltf/include/CesiumGltf/PropertyArrayView.h
@@ -488,14 +488,14 @@ public:
     this->_storage.resize(stringData.size() + offsetData.size());
     std::memcpy(this->_storage.data(), stringData.data(), stringData.size());
     std::memcpy(
-        this->_storage.data() + static_cast<int64_t>(stringData.size()),
+        this->_storage.data() + static_cast<int32_t>(stringData.size()),
         offsetData.data(),
         offsetData.size());
 
     this->_view = PropertyArrayView<std::string_view>(
         std::span<const std::byte>(this->_storage.begin(), stringData.size()),
         std::span<const std::byte>(
-            this->_storage.begin() + static_cast<int64_t>(stringData.size()),
+            this->_storage.begin() + static_cast<int32_t>(stringData.size()),
             offsetData.size()),
         offsetType,
         static_cast<int64_t>(numberOfElements));
@@ -514,7 +514,7 @@ public:
     this->_view = PropertyArrayView<std::string_view>(
         std::span<const std::byte>(this->_storage.begin(), valueSpanSize),
         std::span<const std::byte>(
-            this->_storage.begin() + static_cast<int64_t>(valueSpanSize),
+            this->_storage.begin() + static_cast<int32_t>(valueSpanSize),
             this->_storage.end()),
         rhs._view._stringOffsetType,
         rhs._view.size());
@@ -528,7 +528,7 @@ public:
     this->_view = PropertyArrayView<std::string_view>(
         std::span<const std::byte>(this->_storage.begin(), valueSpanSize),
         std::span<const std::byte>(
-            this->_storage.begin() + static_cast<int64_t>(valueSpanSize),
+            this->_storage.begin() + static_cast<int32_t>(valueSpanSize),
             this->_storage.end()),
         rhs._view._stringOffsetType,
         rhs._view.size());
